+++ src/main/java/com/codeloon/ems/model/InventoryBean.java
@@ -17,8 +17,8 @@ public class InventoryBean {
     private Long itemId;
     private String itemName;
     private Boolean isRefundable;
-    private Long purchasePrice;
-    private Long salesPrice;
+    private Double purchasePrice;
+    private Double salesPrice;
     private Integer orderQuantity;
     private Integer salesQuantity;
     private Integer balanceQuantity;
